Move authentication routes to account.py, resequence models
